Improves handling of invalid Pro licenses
Updates the handling of invalid WP ULike Pro licenses.
The plugin now displays a warning message if an invalid license is detected, but continues to function.
This change ensures that users are notified of the issue without completely disabling the plugin.
Also modifies the message displayed to users with potentially nulled versions of the plugin.