Updates custom template variables and usage
Clarifies available variables and usage within the custom template function.

Improves the example HTML structure by adding a conditional to display
the counter if `$display_counters` is true.

Also, the conditional logic for displaying liked/unliked status is improved
to check the `$status` variable and uses the `formatted_total_likes` for
the counter value.